update ingredient validation

+++ backend/src/recipes/models.py
@@ -59,7 +59,7 @@ class RecipeIngredient(models.Model):
         Recipe, on_delete=models.CASCADE, verbose_name='Рецепт')
     ingredient = models.ForeignKey(
         Ingredient, on_delete=models.CASCADE, verbose_name='Ингредиент')
-    amount = models.FloatField(verbose_name='Количество')
+    amount = models.PositiveIntegerField(verbose_name='Количество')
 
     class Meta:
         constraints = [
